Use the native abbreviation in CurrencyFormatter instead of AbbreviatedFormatter (#535)
* Minor fix

* Use the native abbreviation in CurrencyFormatter instead of AbbreviatedFormatter

* Make abbreviation threshold configurable

* Update PriceViewModel.swift

* Update PriceViewModelTests.swift

* Fixes for currency formatter accuracy

* AssetRateFormatter fix

* String for Double value

* Double parameter instead Decimal in CurrencyFormatter

* CurrencyFormatter tests fix

* CurrencyFormatter calls fix

* ExplorerServiceTests the failing expectation has been disabled.

* Update test

---------

+++ Features/FiatConnect/Sources/ViewModels/FiatQuoteViewModel.swift
@@ -30,7 +30,7 @@ public struct FiatQuoteViewModel {
 
     public var amountText: String {
         switch quote.type {
-        case .buy: formatter.string(decimal: Decimal(quote.cryptoAmount), symbol: asset.symbol)
+        case .buy: formatter.string(double: quote.cryptoAmount, symbol: asset.symbol)
         case .sell: formatter.string(quote.fiatAmount)
         }
     }
